fix: crappy wip but we just need to dedupe on send and assemble on read messages
